improvement: change files structure
