+++ Compte/Shared Components/TapView.swift
@@ -39,38 +39,3 @@ struct TapView<ButtonRepresentation: View>: View {
         .onTapGesture { action?() }
     }
 }
-
-// MARK: - Preview
-struct TapView_Previews: PreviewProvider {
-    static var previews: some View {
-        ZStack {
-            Image(systemName: "person.2.wave.2")
-                .resizable()
-                .scaledToFill()
-            TapView(representation: {
-                Text("Add New Compte")
-            }, buttonFont: { .system(size: 50) })
-            .background(.blue)
-            .clipShape(Capsule())
-            .shadow(radius: 10)
-            .padding()
-        }
-    }
-}
-struct MAXWIDTHTapView_Previews: PreviewProvider {
-    static var previews: some View {
-        ZStack {
-            Image(systemName: "person.2.wave.2")
-                .resizable()
-                .scaledToFill()
-            TapView(buttonFont: {
-                .system(size: 50)
-            }, maxWidth: { .infinity })
-            .background(.blue)
-            .clipShape(Capsule())
-            .shadow(radius: 10)
-            .padding()
-        }
-    }
-}
-
