+++ Compte/Views/BoardList/Components/BoardScrollView.swift
@@ -65,17 +65,3 @@ struct BoardScrollView: View {
         .background(decorator.backgroundColor)
     }
 }
-
-// MARK: - Preview
-struct BoardScrollView_Previews: PreviewProvider {
-    static var previews: some View {
-        var items = Array(repeating: CompteObject.defaultImplementation(), count: 50)
-        NavigationView {
-            BoardScrollView(items: .constant(items)) { id in
-                items.removeAll(where: { $0.id == id })
-            } rename: { _ in
-                // Silent is gold
-            }
-        }
-    }
-}
